Add Redux smart component(container)
