+++ src/components/UnauthenticatedRoute/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
-import { connect } from 'react-redux'
+import { connectAuthen } from '../../containers'
 
 const querystring = (name, url = window.location.href) => {
   name = name.replace(/[[]]/g, "\\$&");
@@ -33,10 +33,4 @@ const UnauthenticatedRoute = ({ component: C, isAuthenticated, props: cProps = {
   );
 };
 
-const mapStateToProps = (state) => ({
-  isAuthenticated: state.authenticate.isAuthenticated
-});
-
-export default connect(
-  mapStateToProps
-)(UnauthenticatedRoute)
+export default connectAuthen(UnauthenticatedRoute)
