+++ src/screens/Home/index.tsx
@@ -3,7 +3,7 @@ import { API } from 'aws-amplify';
 import { PageHeader, ListGroup, ListGroupItem } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
 import "./index.css";
-import { connect } from 'react-redux'
+import { connectAuthen } from '../../containers'
 
 class Home extends Component {
   constructor(props) {
@@ -88,10 +88,4 @@ class Home extends Component {
   }
 }
 
-const mapStateToProps = (state) => ({
-  isAuthenticated: state.authenticate.isAuthenticated
-});
-
-export default connect(
-  mapStateToProps
-)(Home);
+export default connectAuthen(Home);
